Add business constants to new constants ts - businessConstants.ts and remove mapping.ts
