+++ frontend/src/constants/index.ts
@@ -1,6 +1,6 @@
+export * as BIZCONSTANTS from './businessConstants'
 export * as CONSTANTS from './constants'
 export * as CSVHEADERS from './csvHeaders'
 export * as DEFAULTS from './defaults'
-export * as MAPPINGS from './mappings'
 export * as MESSAGE from './message'
 export * as OPTIONS from './options'
