--- frontend/src/constants/mappings.ts
@@ -1,71 +1,0 @@
-// Mapping species code and species names
-export const SPECIES_MAP = {
-  AC: 'Poplar',
-  AT: 'Aspen',
-  B: 'True Fir',
-  BA: 'Amabilis Fir',
-  BG: 'Grand Fir',
-  BL: 'Alpine Fir',
-  CW: 'Western Red Cedar',
-  DR: 'Red Alder',
-  E: 'Birch',
-  EA: 'Alaska Paper Birch',
-  EP: 'Common Paper Birch',
-  FD: 'Douglas Fir',
-  H: 'Hemlock',
-  HM: 'Mountain Hemlock',
-  HW: 'Western Hemlock',
-  L: 'Larch',
-  LA: 'Alpine Larch',
-  LT: 'Tamarack',
-  LW: 'Western Larch',
-  MB: 'Bigleaf Maple',
-  PA: 'Whitebark Pine',
-  PF: 'Limber Pine',
-  PJ: 'Jack Pine',
-  PL: 'Lodgepole Pine',
-  PW: 'Western White Pine',
-  PY: 'Ponderosa (Yellow) Pine',
-  S: 'Spruce',
-  SB: 'Black Spruce',
-  SE: 'Engelmann Spruce',
-  SS: 'Sitka Spruce',
-  SW: 'White Spruce',
-  YC: 'Yellow Cedar',
-}
-
-// Mapping Site Species(Key) and Species Group
-export const SPECIES_GROUP_MAP: { [key: string]: string } = {
-  AC: 'AC',
-  AT: 'AT',
-  B: 'B',
-  BA: 'B',
-  BG: 'B',
-  BL: 'B',
-  CW: 'C',
-  DR: 'D',
-  E: 'E',
-  EA: 'E',
-  EP: 'E',
-  FD: 'F',
-  H: 'H',
-  HM: 'H',
-  HW: 'H',
-  L: 'L',
-  LA: 'L',
-  LT: 'L',
-  LW: 'L',
-  MB: 'MB',
-  PA: 'PA',
-  PF: 'PA',
-  PJ: 'PL',
-  PL: 'PL',
-  PW: 'PW',
-  PY: 'PY',
-  S: 'S',
-  SB: 'S',
-  SE: 'S',
-  SS: 'S',
-  SW: 'S',
-  YC: 'Y',
-}
